[Spark] Move catalog compatibility into shim
Isolate Spark CatalogStorageFormat ABI handling behind a connector compatibility shim.
Clarify the Spark 4.2 CI lane as compile-only while Delta lacks Spark 4.2 artifacts.
